pipe_error_stdout

+++ srcs/unfold/unfold_quote_variable.c
@@ -6,7 +6,7 @@
 /*   By: csakamot <csakamot@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/03 14:59:58 by csakamot          #+#    #+#             */
-/*   Updated: 2023/10/22 13:17:38 by csakamot         ###   ########.fr       */
+/*   Updated: 2023/10/22 17:36:22 by csakamot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ static char	*find_nbr_questoin(t_env *env, char *str, size_t len)
 	result = NULL;
 	if (!ft_strncmp("?", str, len))
 	{
-		result = ft_itoa((int)env->status);
+		result = ft_itoa(env->status);
 		if (!result)
 			exit_malloc_error();
 		return (result);
